Added an incredibly simplistic MVC controller generator.
Use: mvctool classname

Assumes that:

1) Classname is the same as the file, e.g. MainForm is in MainForm.cs and MainForm.Designer.cs
2) the file Xxxxx.Designer.cs is in the *current* directory
3) You are only using standard Winforms controls - if not, edit the source to add extra declarations in. In a future version, this will be a command line option.

How it works - Brute force!! It compiles the designer with a stub class that inherits from System.Windows.Forms.Form, creating an assembly. It then loads this assembly and uses reflection (well... not really, but kind of) to get the info about the exposed controls.

All a little pants, but this is meant for quick starts, rather than correctness!! Ultimately, you get what you pay for, this is FREE!! :-)