This is the initial check in of my old BeOS R5 PDA input driver(with provisional fix for Dano R5.1 input server.)
The code is basically an input server filter driver that interprets data from the user and translates it in to key presses.

This version has NEVER been compiled and is the amalgamation of the PowerPC and Intel R5.03 version with the Dano fix I came up with to work around the major mouse tracking bug in the Input Server.