Version 1 of the RatCow.ComicBook.API. This version will read CBR and CBZ files, but will not handle the INF and will not do any thing spectacular regarding memory... it basically just caches pages as they are opened, but does not ever free the memory. So you will use HUGE amounts of memory if you open a large CBx file and go end to end. It also has a few issues with regards to opening multiple CBx files from the demo app, though that is more likely to do with references being retained.
This is a no thrills first stab at the API, it is not perfect, your mileage will vary.

The third party assemblies NUnrar.dll and Ionic.Zip.*.dll are included because they will aid in getting going and IIRC I used an older version of one or the other libs as I was using 3.5 and the latest uses 4.x as a minimum. The RatCow.MvcFramework.dll and associated binaries are included on a similar tip - but you could also use the latest version built from SVN.