Exploring PDF file manipulation.
This initial check in has a PDF file splitter and recombiner. It takes a standard PDF file and splits the pages in to single items. It then created a "dummy" page and aligns the pages in to booklet order (so if printed on A4 back to back, and the output is folded in the middle, the pages make a "print order" booklet.)

Very basic, no thrills and pretty hardcore. Made to avoid paying a ridiculous ransom to a "pro" app that does the same thing.