Added a System.Windows.Forms.Form helper. This will be extended to enhance/implement the view functionality. Hopefully abstract a lot of the details away.
Altered the ModalSubFormController to remove a compiler warning regarding my reuse of the same template type (and the type was meant to be distinct, so this was an error.)

Added a couple of methods to the UIBridge - "Acknowledge" which is the same as "Alert", except not an error (might clean this all up at some point and make more generic.)

Added a "ProcessSystemMessage" which in WinForms Maps to Application.DoEvents(). This allows me to make sure the main thread isn't being starved of messages.