Made a lot of changes.
o Compiler now has the ability to generate partial methods. This is better for situations where the inheritance model is not wanted.
o Compiler now has options to turn off the listView protection.
o Demo testapp2 now shows how to implement a listView and the use of partial methods. (use the compile35.bat file)
o Cleaned up the code generation a lot.

There's a few more things, but that's most of it.