Some changes:
o added the command line switch "-d", which generates the code in to a "XxxxxController.Designer.cs" file. This then makes the IDE experience more pleasant (the tree groups and folds the files in to one.)
o added the switch "-D". This essentially does the same thing, but it also generates a stub "XxxxxController.cs" file as well if one doesn't exist. This is a shortcut rather than being awfully useful.

In the testapp2:

o reorganised the files in to .\Views and .\Controllers directories. 
o used a new script that generically picks up all of the Xxxxx.Designer.cs files in the .\Views directory and builds/moves a Controller for each to the .\Controllers directory (also handling the stub creation.) This will pick up every file, but skips the ones we create (as they are gone before the next iteration.)
o added the pre-build event in so all of this happens at compile time (never perfect, but it never is with MSBuild.)

What this therefore gives you is the ability to do the same as the previous version, but it also allows better IDE integration. The testapp2 project is a template for configuring a new project.