Changes:
In MvcFramework assembly:

o Added GetSelectedItemOrDefault() method to the ListViewHelper
o Added IDisposable interface implementation to the Basecontroller
o Added more hooks, including ViewClosed, ViewClosing, and hooked Dispose call when View_Disposing is called

In ControllerCreationEngine assembly:

o When compiling the dummy assembly, surfaced the "components" IContainer to the code
o Added in code to handle the StatusStrip special case (it has owned controls in the Items container, rather than Controls)
o Experimental : Added in code to handle Components that don't appear in the Controlls container. This does work, but is not greatly tested.
o Added a check for controls with a null name (this seems to be a oddity of the ProgressBar implementation for StatusStrip.

Altered the testapp2 to test the new code above.

Small code clean-ups and housekeeping.