Added an MVCFormController item template. This creates a new blank Form then creates a Controller with the appropriate name. It does this using the directory format you would expect (.\Controllers\xxx and .\Views\xxx) with the only caveat being that if you don't run it on the *root* of the project, it *will* create the new files in a sub dir of the current project location. I'll look in to fixing it to project root, but not sure that will work. There doesn't seem to be a way to specify "project root". The only other issue is that it will use whatever namspace resolves formt the location you are in... this is not ideal, but it works if you use the project root.
TBH, there's no reason the users can't create controllers and views at different levels, but there's no custom build action yet, so the script that creates controllers will fail to locate the new files. I could make that script recursive, but as it looks for files called "xxx.Designer.cs" and we use that for the controllers too, it'll not really work... The real fix would be to create a custom build action that we can associate. Will look in to that.

I also restricted all item templates to subtype of RatCow... not sure it that actually "works".

This is all still a bit clunky, but it means that the hassle of creating a new controller via the command line compiler is lessened.