Added a new class library full of utils.
Added a new class library that implements UK bank account validation. This is a WIP and it is pretty messy at the moment. Needs major refactoring. It does pass the test suite I currently have in place, this is about 50% of the test cases presented in the example in the documentation you will find included in the /Docs directory.