Some fixes to the ModalSubFormController to remove a compilation warning (Template name collision.)
Start of a port of the MvcFramework to WPF!! The current implementation works almost "out of the box". The WPF version of the base controller has been stripped bare, and does not include any of the Compact Framework code. It is also missing the ModalSubController and the various helpers (as these will mostly come from the RatCow.Utilities namespace.) But the hand translated test app compiles and runs. (I don't claim to be a WPF expert, so I'm sure the little hack I did to the App.xaml.cs and code in the Application_Startup is "bad".) The point is, the same controller with minimal changes works. The major change was the EventArgs becoming RoutedEventArgs, but as I plan to factor out the direct use of EventArgs style constructs, I don't see this as a biggy. Some of the generated code is still commented out, as there seems ot be no analogue to the ListView.

I'm not sure there's a lot of point to all of this. The main thrust was to get the compiler to work with WPF and that is not any where near done!