Updated the code so that the user can now specify custom assemblies... this is a bit of a hack, and I need to integtate this more.
command line: -i=assemblyname1.dll
              -i=assemblyname1.dll,assemblyname2.dll
              -i="assemblyname1.dll,assemblyname2.dll"
              -i="..\..\Project\Libs\assemblyname1.dll,c:\program files\mylocation\assemblyname2.dll"

If the assembly is in the GAC or the current directory (usual discovery rules) one can ignore the path,
but if it is not, one must specify a path to resolve the location!

What I might do is change the compiler to accept as an alternative a CS project and parse out the actual
designer files, plus the associated assemblies (incl hint path). This probably scales better in the long
run. Not quite yet though!!