Added a cool new little extension that wraps up invoking methods on a specific control during a multithreaded operation in to a neat little single call.
Fixed a bug in the Bank validation that allowed an account/sortcode of both zero to fall through and pass as valid. At some point I will have another look at work out if we should just throw out all sortcodes of a less than X length and acc no's of a less than Y, but as I remember it the rules are complicated and so I'm going to need to re-read them before I do that.